feat: implement linked gx fifo buffer

+++ src/graphics/gx_fifo.h
@@ -1,4 +1,4 @@
 #pragma once
 
 #include "cpu/cpu_types.h"
-void gx_write_to_fifo(CPU *cpu, u32 val, u32 size);
+void gx_write_to_fifo(CPU *cpu, u64 val, u32 size);
